test for update function

+++ api/test/Models.cs
@@ -4,8 +4,8 @@ public class Box
 {
     public int Id { get; set; }
     public string Size { get; set; }
-    public int Weight { get; set; }
-    public int Price { get; set; }
+    public float Weight { get; set; }
+    public float Price { get; set; }
     public string Material { get; set; }
     public string Color { get; set; }
     public int Quantity { get; set; }
